move (r)list cmds tests to own module
